+++ assets/javascripts/kitten/components/navigation/link-list/index.js
@@ -89,7 +89,6 @@ export const LinkList = ({
   href,
   ...others
 }) => {
-  
   checkDeprecatedWeights(weight)
 
   return (